+++ docs/src/markdown/about/changelog.md
@@ -2,6 +2,10 @@
 
 ## 1.7
 
+-   **NEW**: Update supported Python versions to Python 3.8 - 3.12.
+-   **NEW**: Use `coloraide` to handle color parsing.
+-   **NEW**: Vendor `hid` and provide fixes for loading the libraries on Windows with Python 3.8+.
+
 ## 1.6
 
 -   **NEW**: Deprecate the timer argument `--type` in favor of `--cmd` which corresponds to how it is stored in the